+++ CHANGELOG.md
@@ -1,6 +1,19 @@
 # CHANGELOG
 
-## 1.1.0
+## 2.0.0 - 2026-02-15
+
+### Changed
+
+- Minimum PHP version 7.
+- Internal: Replace MetaRush\DataMapper with MetaRush\DataAccess.
+
+### Added
+
+- Add setLogger() config method.
+
+## 1.1.0 - 2019-04-14
+
+### Added
 
 - Add flushIp() service method.
 - Auto-expire entries in failCount table.
@@ -8,6 +21,6 @@
 - Auto-expire entries in blockCount table.
 - Add setBlockCountSeconds() config method.
 
-## 1.0.0
+## 1.0.0 - 2019-02-02
 
 - Release first version.